fixed several functionalities

+++ src/main/java/com/mehedi/service/BookService.java
@@ -1,15 +1,11 @@
 package com.mehedi.service;
 import com.mehedi.constatnts.AvailabilityStatus;
-import com.mehedi.dto.BookWithUserDTO;
 import com.mehedi.entity.Book;
-import com.mehedi.entity.BookReservation;
 import com.mehedi.entity.User;
 import com.mehedi.exception.*;
 import com.mehedi.repository.BookRepository;
 import com.mehedi.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -28,21 +24,17 @@ public class BookService {
     private UserRepository userRepository;
 
     public Book createBook(Book book) {
-
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-//        User userNow = userRepository.findByEmail(authentication.getName()).get();
+        User userNow = userRepository.findByEmail(authentication.getName()).get();
 
-        
-
-        if (!authentication.getAuthorities().stream().anyMatch(r -> r.getAuthority().equals("ADMIN"))) {
-            throw new AccessDeniedException("Access is denied. You must have the 'ADMIN' role to create a book.");
+        if(!userNow.getRole().equals(User.Role.ADMIN)) {
+            throw new UnauthorizedUserException("You are not Authorized");
         }
 
         if (book.getTitle() == "" || book.getAuthor() == "") {
             throw new IllegalArgumentException("Title and author cannot be null.");
         }
 
-        // Check for duplicate book by title and author
         if(bookRepository.existsByTitleAndAuthor(book.getTitle(), book.getAuthor())) {
             throw new DuplicateBookException("A book with the same title and author already exists.");
         }
@@ -52,20 +44,32 @@ public class BookService {
     }
 
     public Book updateBook(Long bookId, Book updatedBook) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User userNow = userRepository.findByEmail(authentication.getName()).get();
+
+        if(!userNow.getRole().equals(User.Role.ADMIN)) {
+            throw new UnauthorizedUserException("You are not Authorized");
+        }
+
         Book existingBook = bookRepository.findById(bookId)
                 .orElseThrow(() -> new BookNotFoundException("Book not found with id: " + bookId));
 
-        // Update the properties of the existing book
         existingBook.setTitle(updatedBook.getTitle());
         existingBook.setAuthor(updatedBook.getAuthor());
-        existingBook.setAvailabilityStatus(updatedBook.getAvailabilityStatus());
+//        existingBook.setAvailabilityStatus(updatedBook.getAvailabilityStatus());
 //        existingBook.setDueDate(updatedBook.getDueDate());
 
-        // Save the updated book
         return bookRepository.save(existingBook);
     }
 
     public void deleteBook(Long bookId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User userNow = userRepository.findByEmail(authentication.getName()).get();
+
+        if(!userNow.getRole().equals(User.Role.ADMIN)) {
+            throw new UnauthorizedUserException("You are not Authorized");
+        }
+
         Book existingBook = bookRepository.findById(bookId)
                 .orElseThrow(() -> new BookNotFoundException("Book not found with id: " + bookId));
         bookRepository.delete(existingBook);
